modularized owner group dropdown selector
