+++ components/compose-web-app/modules/library.js
@@ -25,7 +25,7 @@ import { LibraryHistory } from "./library-history.js";
 import { LibraryEditInterfaces } from "./library-interfaces.js";
 import { LibraryEditSimple } from "./library-simple.js";
 import { TabSheet } from "./tabsheet.js";
-import { FormLayout, LayoutRow, SetupTable, TextArea } from "./util.js";
+import { FormLayout, LayoutRow, SetupTable, TextArea, OwnerGroupSelector } from "./util.js";
 
 export async function BuildLibraryTable() {
     const response = await fetch('/compose/v1alpha1/library/blocks');
@@ -113,42 +113,7 @@ async function BlockForm(blockTypes, interfaceRoles) {
     composite.textContent = 'Composite';
     bodySelector.appendChild(composite);
 
-    // create owner group selector dropdown
-    let ownerGroupSelector = document.createElement('select');
-    ownerGroupSelector.id = 'ownerGroupSelector';
-
-    try {
-        const ownerGroupResult = await fetch('/api/v1alpha1/user/groups');
-        if (ownerGroupResult.ok) {
-            const ownerGroupList = await ownerGroupResult.json();
-            
-            // Add a default/empty option
-            let defaultOption = document.createElement('option');
-            defaultOption.textContent = '-- Select a group --';
-            defaultOption.value = '';
-            ownerGroupSelector.appendChild(defaultOption);
-            
-            // Add user's groups
-            for (const ownerGroup of ownerGroupList) {
-                let option = document.createElement('option');
-                option.textContent = ownerGroup.name;
-                option.value = ownerGroup.id;
-                ownerGroupSelector.appendChild(option);
-            }
-        } else {
-            // Handle error case
-            let errorOption = document.createElement('option');
-            errorOption.textContent = 'Error loading groups';
-            errorOption.disabled = true;
-            ownerGroupSelector.appendChild(errorOption);
-        }
-    } catch (error) {
-        console.error('Error fetching user groups:', error);
-        let errorOption = document.createElement('option');
-        errorOption.textContent = 'Error loading groups';
-        errorOption.disabled = true;
-        ownerGroupSelector.appendChild(errorOption);
-    }
+    const ownerGroupSelector = await OwnerGroupSelector();
 
     const form = await FormLayout(
         //
