+++ components/compose-web-app/modules/util.js
@@ -362,3 +362,43 @@ export function MultiSelectWithCheckbox(items) {
 
     return layout;
 }
+
+export async function OwnerGroupSelector() {
+    let ownerGroupSelector = document.createElement('select');
+    ownerGroupSelector.id = 'ownerGroupSelector';
+
+    try {
+        const ownerGroupResult = await fetch('/api/v1alpha1/user/groups');
+        if (ownerGroupResult.ok) {
+            const ownerGroupList = await ownerGroupResult.json();
+            
+            // Add a default/empty option
+            let defaultOption = document.createElement('option');
+            defaultOption.textContent = '-- Select a group --';
+            defaultOption.value = '';
+            ownerGroupSelector.appendChild(defaultOption);
+            
+            // Add user's groups
+            for (const ownerGroup of ownerGroupList) {
+                let option = document.createElement('option');
+                option.textContent = ownerGroup.name;
+                option.value = ownerGroup.id;
+                ownerGroupSelector.appendChild(option);
+            }
+        } else {
+            // Handle error case
+            let errorOption = document.createElement('option');
+            errorOption.textContent = 'Error loading groups';
+            errorOption.disabled = true;
+            ownerGroupSelector.appendChild(errorOption);
+        }
+    } catch (error) {
+        console.error('Error fetching user groups:', error);
+        let errorOption = document.createElement('option');
+        errorOption.textContent = 'Error loading groups';
+        errorOption.disabled = true;
+        ownerGroupSelector.appendChild(errorOption);
+    }
+
+    return ownerGroupSelector;
+}
